+++ helpers/logs.py
@@ -220,7 +220,7 @@ def parse_status(steamapi, line, regex, cache, check_private, max_guess_iteratio
                 while created is None:
                     if iterations > max_guess_iterations:
                         return float('inf'), r.group('user_id'), r.group('name'), 'Max guess iterations reached', \
-                               False, None, None
+                               False, None, None, None
                     uuid_working += 1
                     iterations += 1
                     p = steamapi.call('ISteamUser.GetPlayerSummaries', steamids=uuid_working)['response']['players']
@@ -231,22 +231,30 @@ def parse_status(steamapi, line, regex, cache, check_private, max_guess_iteratio
                     except KeyError:
                         continue
             else:
-                return float('inf'), r.group('user_id'), r.group('name'), 'Guessing disabled', False, None, None
+                return float('inf'), r.group('user_id'), r.group('name'), 'Guessing disabled', False, None, None, None
         cache[steam_id] = created, approximate
     else:
         created, approximate = cache[steam_id]
 
     td = timedelta(seconds=time() - created)
     td2 = None
+    lvl = None
 
-    if check_csgo_playtime and not approximate:
+    if not approximate:
         try:
-            p = steamapi.call('ISteamUserStats.GetUserStatsForGame', steamid=sid.as_64, appid=730)
+            p = steamapi.call('IPlayerService.GetSteamLevel', steamid=sid.as_64)
         except HTTPError:
             pass
         else:
-            td2 = timedelta(seconds=next((i for i in p['playerstats']['stats'] if i['name'] == 'total_time_played'),
-                                         None)['value'])
+            lvl = p['response']['player_level']
+        if check_csgo_playtime:
+            try:
+                p = steamapi.call('ISteamUserStats.GetUserStatsForGame', steamid=sid.as_64, appid=730)
+            except HTTPError:
+                pass
+            else:
+                td2 = timedelta(seconds=next((i for i in p['playerstats']['stats'] if i['name'] == 'total_time_played'),
+                                             None)['value'])
 
     server_playtime = None
 
@@ -257,5 +265,5 @@ def parse_status(steamapi, line, regex, cache, check_private, max_guess_iteratio
             server_playtime = "Invalid Server"
 
     return created, r.group('user_id'), r.group('name'), human_readable.precise_delta(
-        td, suppress=find_human_suppress(td)), approximate, human_readable.precise_delta(
+        td, suppress=find_human_suppress(td)), approximate, lvl, human_readable.precise_delta(
         td2, suppress=find_human_suppress(td2)) if td2 is not None else None, server_playtime